+++ SistemaVentas/src/main/java/com/ventasx/SistemaVentas/Controller/ProductController.java
@@ -1,7 +1,9 @@
 package com.ventasx.SistemaVentas.Controller;
 
 import com.ventasx.SistemaVentas.Controller.Dto.ProductDto;
+import com.ventasx.SistemaVentas.Controller.Dto.SuccessMessageDto;
 import com.ventasx.SistemaVentas.Controller.Mapper.MapperBetweenDtoAndEntity;
+import com.ventasx.SistemaVentas.Exception.ResourceNotFound;
 import com.ventasx.SistemaVentas.Persistence.Entity.Product;
 import com.ventasx.SistemaVentas.Service.IProductService;
 import jakarta.validation.Valid;
@@ -13,7 +15,10 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
@@ -48,41 +53,77 @@ public class ProductController extends MapperBetweenDtoAndEntity<ProductDto, Pro
     }
 
     @GetMapping
-    public ResponseEntity<List<ProductDto>> listAllData() throws Exception{
-        List<ProductDto> listDataDto =  service.getAll().stream().map(this::mapFromEntityToDto).toList();
+    public ResponseEntity<List<ProductDto>> listAllData() {
+        List<ProductDto> listDataDto =  service.findAllByIsEnabledTrueOrderByCreationDateDesc().stream().map(this::mapFromEntityToDto).toList();
+        return new ResponseEntity<>(listDataDto, HttpStatus.OK);
+    }
+
+    @GetMapping("/disabled")
+    public ResponseEntity<List<ProductDto>> listAllDataDisabled() {
+        List<ProductDto> listDataDto =  service.findAllByIsEnabledFalseOrderByCreationDateDesc().stream().map(this::mapFromEntityToDto).toList();
         return new ResponseEntity<>(listDataDto, HttpStatus.OK);
     }
 
     @GetMapping("/pagination")
-    public ResponseEntity<Page<ProductDto>> listAllDataPagination(@PageableDefault(sort = "creationDate", direction = Sort.Direction.DESC) Pageable pageable) throws Exception {
-        Page<ProductDto> dataPagination = service.getAllPagination(pageable).map(this::mapFromEntityToDto);
+    public ResponseEntity<Page<ProductDto>> listAllDataEnabledPagination(@PageableDefault(sort = "creationDate", direction = Sort.Direction.DESC) Pageable pageable) {
+        Page<ProductDto> dataPagination = service.findAllByIsEnabledTrueOrderByCreationDateDescPagination(pageable).map(this::mapFromEntityToDto);
+        return new ResponseEntity<>(dataPagination, HttpStatus.OK);
+    }
+
+    @GetMapping("/pagination-disabled")
+    public ResponseEntity<Page<ProductDto>> listAllDataNotEnabledPagination(@PageableDefault(sort = "creationDate", direction = Sort.Direction.DESC) Pageable pageable) {
+        Page<ProductDto> dataPagination = service.findAllByIsEnabledFalseOrderByCreationDateDescPagination(pageable).map(this::mapFromEntityToDto);
         return new ResponseEntity<>(dataPagination, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<ProductDto> listDataById(@PathVariable("id") Long id) throws Exception{
         Product data = service.getById(id);
-//        if (data == null) {
-//            throw new ResourceNotFound("Paciente", "id", pacienteDto.getPaciente_id());
-//        }
+        if (data == null){
+            throw new ResourceNotFound("Producto", "id", String.valueOf(id));
+        }
         return new ResponseEntity<>(mapFromEntityToDto(data), HttpStatus.OK);
     }
 
     @PutMapping()
     public ResponseEntity<ProductDto> updateData(@RequestBody @Valid ProductDto dto) throws Exception {
-        Product data = service.getById(dto.getId());
-//        if (paciente == null) {
-//            throw new ResourceNotFound("Paciente", "id", pacienteDto.getPaciente_id());
-//        }
-        return new ResponseEntity<>(mapFromEntityToDto(service.update(data)), HttpStatus.OK);
+        if (service.getById(dto.getId()) == null){
+            throw new ResourceNotFound("Producto", "id", String.valueOf(dto.getId()));
+        }
+        return new ResponseEntity<>(mapFromEntityToDto(service.update(mapFromDtoRequestToEntity(dto))), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteData(@PathVariable(value = "id") Long id) throws Exception {
-        Product data = service.getById(id);
-//        if (paciente == null) {
-//            throw new ResourceNotFound("Paciente", "id", paciente_id);
-//        }
+    public ResponseEntity<SuccessMessageDto> deleteData(@PathVariable(value = "id") Long id) throws Exception {
+        if (service.getById(id) == null){
+            throw new ResourceNotFound("Producto", "id", String.valueOf(id));
+        }
         service.delete(id);
+        return new ResponseEntity<>(SuccessMessageDto.builder().statusCode(HttpStatus.OK.value()).timestamp(LocalDateTime.now())
+                .message("Registro eliminado exitosamente.").build(), HttpStatus.OK);
+    }
+
+    @PutMapping("/disable")
+    public ResponseEntity<SuccessMessageDto> disableData(@RequestBody List<Long> listId) throws Exception {
+        for (Long id: listId){
+            if (service.getById(id) == null){
+                throw new ResourceNotFound("Producto", "id", String.valueOf(id));
+            }
+            service.disableProductsById(id);
+        }
+        return new ResponseEntity<>(SuccessMessageDto.builder().statusCode(HttpStatus.OK.value()).timestamp(LocalDateTime.now())
+                .message("Registros deshabilitados exitosamente.").build(), HttpStatus.OK);
+    }
+
+    @PutMapping("/enable")
+    public ResponseEntity<SuccessMessageDto> enableData(@RequestBody List<Long> listId) throws Exception {
+        for (Long id: listId){
+            if (service.getById(id) == null){
+                throw new ResourceNotFound("Producto", "id", String.valueOf(id));
+            }
+            service.enableProductById(id);
+        }
+        return new ResponseEntity<>(SuccessMessageDto.builder().statusCode(HttpStatus.OK.value()).timestamp(LocalDateTime.now())
+                .message("Registros habilitados exitosamente.").build(), HttpStatus.OK);
     }
 }
